requirements.txt
django==2.1.7
djangorestframework==3.9.2
xxhash==1.3.0
jsonschema==3.0.1
psycopg2-binary==2.8.1
gunicorn==19.9.0
